Fix skip condition

+++ stdlib/database/jdbc/src/test/java/org/ballerinalang/jdbc/execute/ParamsExecuteTest.java
@@ -28,6 +28,7 @@ import org.ballerinalang.test.util.BCompileUtil;
 import org.ballerinalang.test.util.BRunUtil;
 import org.ballerinalang.test.util.CompileResult;
 import org.testng.Assert;
+import org.testng.SkipException;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
@@ -49,6 +50,10 @@ public class ParamsExecuteTest {
 
     @BeforeClass
     public void setup() throws SQLException {
+        // Temporary solution for https://github.com/ballerina-platform/ballerina-lang/issues/24227
+        if (System.getProperty("os.name").toLowerCase().contains("win")) {
+            throw new SkipException("Skip test class in Windows");
+        }
         result = BCompileUtil.compile(SQLDBUtils.getBalFilesDir("execute",
                 "execute-params-query-test.bal"));
         SQLDBUtils.deleteFiles(new File(SQLDBUtils.DB_DIR), DB_NAME);
